+++ docs/Changelog.md
@@ -4,6 +4,7 @@
 * Change snackbar builder
 * Change addBundle to withArguments to match ANKO
 * Created KauIItem to replace AbstractItem
+* Created permission manager and permission constants
 
 ## v1.4
 * Added about activities